requirements.txt
flask-cors==3.0.2
Flask
logzero 
praw
gmusicapi
google-cloud-language
python-dotenv